Add files via upload
The login feature is the main emphasis of this initial phase of the Chat Application project.  In order to access the chat system, users must first submit a valid username and password in the login module.  It has an easy-to-use interface for user authentication, error handling for inaccurate information, and rudimentary form validation to stop blank entries.  A default test account is supplied for easy access, and the module, which was constructed with Java, presently stores credentials locally.  To get started, clone the repository, go to the login folder, and use the Java Login command, which is the right command for your language.  Secure password hashing, user registration, and database integration are planned future improvements.